Reorganize to match stripes-cli blueprint
