+++ src/ConnectedInstance/ApolloConnectedInstance.js
