+++ src/ConnectedInstance/ConnectedInstance.js
