+++ src/ConnectedInstance/StripesConnectedInstance.js
