+++ src/ConnectedInstance/index.js
