+++ src/edit/publicationFrequencyFields.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import TextField from '@folio/stripes-components/lib/TextField';
-import RepeatableField from '../src/components/RepeatableField';
+import RepeatableField from '../components/RepeatableField';
 
 const PublicationFrequencyFields = ({ formatMsg }) => (
   <RepeatableField
