+++ src/settings/ContributorTypesSettings.js
