+++ src/settings/ResourceTypesSettings.js
